:art: Refactor pseudo pext benchmarks to separate algorithms
